% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alkhwarizm.R
\docType{package}
\name{alkhwarizm}
\alias{alkhwarizm}
\title{alkhwarizm: A package for executing popular 'Euclidean' and 'Dijkstra' algorithms.}
\description{
The alkhwarizm package provides two categories of important functions:
euclidean, dijkstra
}
