% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find 'Greatest Common Divisor'(GCD) for two given numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A positive integer.}

\item{b}{A positive integer.}
}
\value{
The GCD of \code{a} and \code{b}.
}
\description{
Find 'Greatest Common Divisor'(GCD) for two given numbers
}
\examples{
euclidean(10,100)
euclidean(123612, 13892347912)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
